requirements.txt
# Core Flask dependencies
Flask==3.1.2
Flask_Cors==4.0.0
flask_limiter==3.12
Werkzeug==3.1.3
Jinja2==3.1.6

# AI and ML dependencies
groq==0.31.1
python-dotenv==1.1.1

# Document processing
pdfplumber==0.10.3
PyPDF2==3.0.1
python_docx==0.8.11
mammoth==1.6.0
python_magic==0.4.27

# Image processing
Pillow==11.3.0

# HTTP and networking
requests==2.31.0
urllib3==2.0.7

# Security and cryptography
cryptography==45.0.7
pyOpenSSL==25.1.0

# Utilities
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
importlib_metadata==8.0.0
lockfile==0.12.2
protobuf==6.32.0
truststore==0.10.4

# Development and testing
mock==5.2.0
docutils==0.22
railroad==0.5.0
Sphinx==8.2.3
tornado==6.5.1

# Jupyter and interactive
ipython==8.12.3
ipywidgets==8.1.7
keyring==25.6.0
thread==2.0.5
HTMLParser==0.0.2
xmlrpclib==1.0.1
urllib3_secure_extra==0.1.0
